apps/auth-service/requirements.txt
boto3
uuid==1.30
pyjwt
bcrypt
